fix(redis): Decode redis responses to strings
The redis.py docs say:

> By default, all responses are returned as bytes in Python 3 and str
  in Python 2. The user is responsible for decoding to Python 3
  strings or Python 2 unicode objects.

> If all string responses from a client should be decoded, the user
  can specify decode_responses=True to Redis.__init__. In this case,
  any Redis command that returns a string type will be decoded with
  the encoding specified.

In the py2 world, it didn't really matter if you did this properly,
but of course, py3 is strict like a beast, and you gotta get this
right (good!).

Also adds a test for the profile page's API page, which currently
fails if an anonymous user makes a query and then you try to load it.

Fixes: COURTLISTENER-JC
Fixes: #1449