fix(docker): Stop spawning judge pics and seal rookery
Note that the functionality here varies a bit between swarm
deployments and regular ones. In a swarm, the default restart policy
is "any." That actually isn't defined in the documentation and is
impossible to google, so I honestly don't know what it's supposed to
do, but it was restarting things A LOT in the swarm. Instead I've
changed it to "none".

However!

1. The default outside of a swarm, when you're just using docker
   compose to deploy, is "no". So we only need to override the default
   when we deploy to swarms,

2. The "deploy" key is ignored when you don't do a swarm.

Putting this together, I've added the deploy key everywhere. It's not
needed except in the task-server compose file, because the others
aren't deployed as a swarm and have the default we want ("no").

Fixes: #1472