fix(tasks): Do not rely on py2 exceptions scope
In py3, for technical reasons, they made the language a bit worse by
narrowing the scope of exceptions. In the past, this worked:

    try:
        1/0
    except ZeroDivisionError as exc:
        pass
    print(exc)

Now that fails b/c exc is out of scope when you get to the print
statement. PyCharm highlights these errors now, thanksfully, so I did
it report on all of them it could find so I could try to wipe them all
out. Most seem fine, but I did find these two that needed fixing.

Fixes: #1474
Fixes: COURTLISTENER-MP