fix(citations): Add support by modification date
This removes a old comment warning not to do a feature, and then adds
that feature. The worry expressed in the comment was that if you query
by date modified and then modify documents, you'll wind up with an
endlessly running loop. At the time, I believe, that was true.

That has now changed though, because Django now uses server-side
cursors via the `iterator()` method on the queryset. What this does is
open a transaction in the DB, query the results, and then disk them
out to us bit by bit. Our old way used to fetch chunks of rows at a
time, keeping track of our position as we paged through the result
set.

A few docs I've read report that when you open a cursor, the results
are fetched serverside. I don't think that's how cursors work because
the DB would have to keep huge result sets in memory all the time.
What I think it does instead is uses MVCC to mark the moment that the
cursor is opened, and to keep track of rows that are changed after
that moment. Then, when those rows are fetched, it can give the
earlier version of those rows rather than the current one, until the
cursor is closed again.

Anyway, the big picture is that by using a cursor here, we get a
snapshot of the data at that moment. Thus, our changes to the
date_modified field in a Celery task won't affect the result set in
the cursor.