fix(scrapers): Stop catching and logging errors; raise instead
Right now fixing erorrs in our scrapers is hard because of the way we
log them instead of raising them. This change raises them properly,
which is going to crash a lot of crawlers in the short term, but which
will make them easier to debug in the long term.