fix(nginx): Remove X-Frame-Option header completely
I've spent some time looking at this, but I don't think there's a
simple way to override this header at the request of the Django view.
I think I could maybe do it by creating a new location block for
/embed/ or whatever, but I don't like adding a new location block just
for the one header.

Other ideas:

 - Use a middleware to do it. Could work. The middleware could set
   the header on all responses, and could have a trick for exepting
   certain views. Could work, but a PITA and it wouldn't add the
   header to static files, so we'd still need to do that somehow.

 - Just set the header in Django. I tried this, but doing so results
   in a duplicate header on the client.

So maybe there's a solution in Django that's kind of elegant, and
maybe there's one in nginx that's workable, but it's all kind of a lot
just to accomplish a very small security enhancement. Thus: Let's
remove the header, restore functionality, and maybe someday we'll add
it back.

Fixes: #1489 (again)