fix(sitemaps): Reduce number of sitemaps
It turns out that most of the problem comes from sorting. By using a
large limit, you reduce the number of pages that might need to be
generated, and thus reduce the number of times the server might need
to sort results. The tradeoff is that the sitemaps generated are
truly huge.