fix(sitemaps): Add longer and more granular cache
This is a lot of code to accomplish something pretty simple. All this
does is copy the code from Django's included sitemap view while adding
a small cache to it. Prior to this, we used the cache_page decorator,
which is fine, sort of, except that it caches *all* of the pages,
including our last one, which under normal circumstances is growing.

We could just cache that page for two weeks and eventually, yes, the
page would be correct. But I want to make the cache last for a truly
long time — 180 days as of this commit — and so having a bad sitemap
cached that long really wouldn't be great.

So in this version, we set the cache length based on whether the
sitemap is complete. If it is, great: 180 days. If not, fine: 1 day.

This version also improves our cache by only caching the URLs instead
of also caching the serialized XML (as we all know, XML is a verbose
mess, so not caching all that verbosity will shrink our cache usage).