feat(iquery): Add a script for gathering all iquery results
This is a humble script, but it's going to do a big thing. It will
begin scraping all iquery pages starting at number 1 and then going up
until there are no more hits. Long term, hopefully, this script will
run as a daemon and make sure that we have some information about
every case in the system at all times.