fix(robots): Don't include recent cases in robots.txt
Generally, there's no need to put them here, if they were blocked at
the time of their creation (which is common for bankruptcy cases). The
only reason we put recently blocked cases on the list is to get them
off search results as quickly as possible. New cases wouldn't be on
search results.