fix(law): Don't block access to the law
With generous support from the Samuelson Clinic at UC Berkeley, we are
proud to have fought to keep this case available to the public. During
the course of fighting to keep this content available, we briefly
removed it in order to comply with a sealing order issued by Judge
Preska.

We are pleased today to have convinced the Court that they cannot seal
this case, and as such, we are putting it back in its rightful home.