fix(tasks): Eliminate race condition
Thumbnails are generated the first time a FD is saved, by the save
method. Alas, that method did not have a select for update, so it was
possible — indeed likely — that this race condition could happen:

    1. Thumbnail task starts, and selects the FD.
    2. Extraction of the FD finishes, saves stuff to the DB.
    3. Thumbnail task finishes, saves to the DB.

If the above happens, step 2 gets overwritten by the content SELECT'ed
in step 1, and that's bad.

This uses select_for_update, a method we should be using much more
frequently, to lock the row until the thumbnail task completes.