feat(auth): Upgrade ConfirmedEmailAuthenticationForm
This could have been a one-line change to add the request parameter as
the first positional argument to the authenticate function, but while
doing that I noticed that Django now has a hook for
`confirm_login_allowed` that makes this more surgical and tidier.
Great! Now we use that instead and we don't use the authenticate
function at all anymore.