fix(donations): Add support for Stripe Invoice payments
Some folks can't use our donate page, but can pay through Stripe
Invoices (Amex, no American address, etc.). For those folks, when we
get the callback, it was crashing. This fixes this. Unfortunately,
the callbacks for these folks don't have *any* of their information,
so we create a user called "nobody@stripe.com" to hold their
donations.

I also did a very big refactor while working on this, so this code is
a bit broken. The next commit lands the refactor.