refactor(donations): Pull code into functions
This code is always a rat's nest, but this should make it a bit
better. General idea is to pull things out of the main function and
put it in smaller, more manageable ones:

 - update_donation_for_event: Updates the donation object according
   to the event itself.
 - send_thank_you_if_needed: Sends the thank you (using the methods we
   have for that), but only if needed (e.g., not on a Xero payment).
 - get_donation_with_retries: Does the donation lookup retry maneuver.

This also uses type hints, good doc strings, and f-strings where
possible to make the code a lot leaner and tidier.