feat(settings): Allow failing judge pics import
In mypy, when you use the django-stubs project, and configure it using
the mypy.ini file, it uses your settings file as part of the
inspection. For that to work, you have to have the dependencies
available, which we only do in a docker environment. So if you want
mypy to work without docker being up and running (like in CI, say),
you have to allow the import to fail.

Note that we use a random passphrase on the path. The idea here is to
protect against *other* reasons the import might fail. Imagine that a
typo slips through into the import statement, so it says:

    from juuuuudge_pics import judge_root

That could get deployed to prod, and if the judge_root were set to,
say, "/", that'd give more access to our root file system than we
should. The random passphrase prevents this, all but guaranteeing that
any errant import errors in prod will have no access to anything.