fix(s3): Allow custom domains in our storage class
In #1033, we moved audio files to S3, but after doing so the URLs were
wrong when you called url() on a FileField. This was partly fixed in
the prior commit, which fixed the custom domain in the settings, but
this commit was also necessary to allow the setting to take hold.

I'm a bit concerned about this b/c I worry that we may have issues
with uploads to S3 if CF is always in the path, but we'll see. I
consulted with @flooie, and he's not sure why the custom_domain was
disabled (it's been awhile). I don't think this will cause breakage,
but it's the right fix for the moment, so I'm going for it.