feat(feeds): Update feeds to use S3
This removes the enclosure length functionality. We don't keep
this around and once content is in S3 it'll mean downloading
each item to get a measurement of the length and put it in the
RSS feed. Since we don't want to do that, pull the
functionality. The feeds will be fine, just not as optimized.