feat(models): Upgrade to AWS classes
In the historical migrations, I had to remove the incrementing file
storage since the class is now deleted from the codebase. Luckily just
removing it reverts the mdoels to the default storage class, and then
our latest migrations just override that.

The result is a noop migration at the end. Precisely what we're
looking for.