feat(s3): New view with tests for redirection
This will get traffic from S3 when open graph bots ask for a file.
Instead of giving it to them from S3, the response will come here and
we will generate and serve open graph data. It's a beautiful hack.

If we can't find the document here for some reason we abort and send
them back to S3, but with the open graph stuff disabled.