fix(urls): Correctly catch bad URLs
This goes to show that messing with security-critical code is a nasty
business and often a mistake. Luckily, we have tests, but what
happened here is that I tweaked the code to separate out checks from
redirection. In so doing, I removed pulled the quoting code into the
checks, but took it out of the redirection part. So previously, we'd
quote a URL before redirecting it, now we didn't.

To fix this, I copied the fix from Python's standard lib that you can
see here:

    https://github.com/python/cpython/pull/25725/

Instead of quoting it, we just make sure it doesn't have newlines. If
it does, we bail.