fix(templates): Normalize on the static tag
This overhauls the way we access assets inside our templates. The way
we have done it so far is pretty complex, inconsistent, and brittle
and Django has a much better way. In addition, by doing it this way,
we prepare the code to use the ManifestStaticFilesStorage class,
which will make and load convenient hashes of our files for us.

To accomplish the overhaul, this:

 - Evaporates the STATIC_PREFIX variable that was getting set in
   base.html and then read all over the place;

 - Removes the version GET params we were adding to some assets to
   bust caches; and

 - Stops using STATIC_URL, replacing it with the static template.