fix(js): Remove hot reloading
Unfortunately, I'm really not sure what's going on here, but when
these are installed, I get dumb errors in the console akin to those
discussed here:

https://github.com/pmmmwh/react-refresh-webpack-plugin/issues/302

After spending quite a while wrestling with it and not really
understanding it, I gave up and yanked the functionality. We can build
it back later, but for now the goal is to land a feature, not
necessarily to land a good developer experience (sigh!). Perhaps we
can call in an expert to help.