fix(js): Remove hashed file names
This change is in pusuit of the approach explained on this page:

https://pascalw.me/blog/2020/04/19/webpack-django.html

The general idea is to *not* hash things with Webpack and to instead
let Django's ManifestStaticFilesStorage class do the hashing.

This (and the previous commit) also remove the Compression plugin from
webpack, since it isn't something nginx is even set up to use at the
moment. See:

https://slashgear.github.io/why-you-should-use-compression-webpack-plugin/

So, I think this is a good step. Also, it works!