feat(files): Configure ManifestStaticFilesStorage
There's a lot of spooky action at a distance on this one, but I think
this summarizes it well. Configuring this variable will create and
serve static files with nice hashes in their names, when:

1. The template uses {% static %}. (See earlier commit doing that
across all our templates.)

2. DEBUG is set to False.

And to generate those files so they actually exist, you have to run
collectstatic with DEBUG=False and the storage setting in place.

This also hides the setting during tests, because, well, I just work
here. The docs tell you to.