fix(celery): Simplify CeleryThrottle
I've noticed that for quite awhile the celery throttle didn't seem to
work quite right. Over the years, I put in a couple fixes that made
the code longer and more complex, and when I went to fix it yesterday,
I decided that I didn't like it anymore and that I should just
reimplement it.

This new implementation largely does what the old one did, but in a
simpler way. Instead of trying to be clever about measuring the
average length of task processing on the queue and then working with
that, this new version just has a flat (but configurable) polling
interval that it uses when the queue is swamped.

The new algorithm simply:

1. Loads up the queue up to 2× self.min.
2. Measures if the length of the queue is less than self.min. If so,
   loads up the queue so it's back to self.max.
3. If, on the other hand, the queue is longer than self.min, it waits
   a `poll_interval` amount of time and measures again, and again, and
   again, until it's below self.min. When that finally happens, it
   gets loaded up to self.max again.

That's it! Much simpler. Relies on polling, but in a relatively
harmless way (polling every few seconds is *fine*).

Fixes: #1715