fix(celery): Remove race condition
Add a check to make sure that if the incr creates the key, we give it
a ttl. Without this, sometimes the key can expire between when we get
it and when we incr it. If that happens, the incr creates the key, and
it lacks a TTL, which sucks. This should fix that.