fix(users): Typo in account deletion function
This was just a typo, but obviously we should have tests that catch
such typos, so I added that too. In doing so, I caught a separate bug
that our mailchimp code might not have been working due to the logout
command just preceding it, so I changed the order there. I also
removed the blanket `except` statement since it only served to trick
the user when failures occured. Better to surface them.

Fixes: COURTLISTENER-1K5