feat(celery): Add a scheduler to the throttle
I spent a week figuring out how to do this and researching alternative
approaches. I researched giving up Celery (again), moving things to
AWS Batch, etc.

Much of my trouble was caused by wanting to build an exponential
backoff algorithm into each task so they could be retried slower and
slower. It turns out that what you really want to do is build a
backoff algorithm for the entire task type, not for each task. This
makes sense: Why give each task its own backoff when a centralized
location is controlling their rate anyway?

The solution, it finally turns out, is to create a scheduler for the
tasks so that when they are throttled they are told, "Look, there are
X tasks in front of you. Those tasks will take Y seconds. Try back
after that."

This commit implements that algorithm. I kind of hate how complex it
wound up, and I predict bugs later, but so far in testing it works
well. Better than the status quo anyway, which currently has a massive
runaway problem that leads to a huge task queue that's constantly
bouncing around between throttling a task and retrying it, over and
over ad naseum.