fix(harvard_opinions): Speed up harvard importer
We introduced a bug in the speed of our importer
when we added an if any[x,y,z,a,b] in match_based_text
becuase this code made xyza and b all get called
everytime.  Instead we just needed an
if or or or or or statement that bails when sometime
fails.

We should see a signficant speed up.  4x to 10x
I believe.