fix(tests): Always load the migration and tweak query count
This was going to be complicated to fix another way. It'd be
nice not to load data during tests, but there aren't that many
governors, so it's probably fine to do so. By loading the
fixture during tests, we can be sure that the tests run the
same whether this is being run as a test or on a db that's
already migrated.