fix(rss): Disable retries
This is a second attempt at fixing the issue resolved by:

369a50837125290687feb3cd9a1cfc0ce5c3f337

The message for that one said:

> This task looks for changes to the RSS feeds, and currently has a
> retry of five times. Because it also has a lengthy timeout, it can
> really tie up a celery workers when a site goes down. What's worse
> is that trying five times, each with a five minute wait means that
> there's a good chance you'll still be trying to retry the *last*
> task when the next scrape of the site comes along (I think). So...if
> a site goes down, or as in the case today, multiple ones do, this
> can totally tie up Celery.

That's a pretty good summary, but today I saw that COB's crawler was
down and that for some reason it was still getting retried every five
seconds. I think that might be because max_retries=0 doesn't work the
way I think/thought it does, but in any case, if we're disabling
retries, let's not have code that can do retries. This commit removes
that code.