refactor(scrapers.tasks): Remove tax specific code
Tax court no longer reparses its code for docket
number.  This was implemented because it was
needed.  This is no longer the case so it has
been removed from extraction.

The tests associated have also been yanked. But I
left the method in tasks because I think it
will be necessary in the near-ish future.