fix(poetry): Tighten python version specification
SciPy's latest version (which will be installed in a subsequent commit) requires a version bound of > 3.8 and < 3.11. The previous version bound of ^3.8 translates to > 3.8 and < 4.0. *This is equivalent in practice* because there is no such thing as Python 3.11, but we need to make this explicit because poetry is not smart enough to figure it out and throws an error when attempting to install scipy