fix(dei): Don't use harmful language
Django 3.2 got rid of harmful language in a few places. We switched
away from using 'master' for the main branch in git, and this removes
still more insensitive language.

I think allowlist and denylist are a bit clunkier than their
predecessors, but we'll get used to them, and they're what django
used, so it's fine.