fix(deps): Yank dependency on pypdf2
In the following issues and PRs, PyPDF2 users have identified and
offered to fix a weird bug where the library overrides normal logging
in a way that breaks when you log certain things:

https://github.com/mstamy2/PyPDF2/issues/67
https://github.com/mstamy2/PyPDF2/pull/641
https://github.com/mstamy2/PyPDF2/pull/452
https://github.com/mstamy2/PyPDF2/pull/243

Not great, but there was probably a reason. Unfortunately, the
maintainers aren't merging any of the fixes people have provided
over the years, and when I upgraded to Python 3.10 one of our tests
changed in a way that it triggered this bug.

So, since the maintainers don't seem inclined to fix their own
package, this commit yanks it from CourtListener. It's good timing,
really, since we now have the microservice available, but it was
disappointing to see bugs and PRs related to this going back to 2014.
Most of the fixes are one or two-line PRs too. Bummer.