fix(settings): Fix tests
It turns out that environ, though a very good package, does a bit of a
funny thing. Instead of relying on pathlib to do path stuff, it ships
its own pathlib-alike. I'm guessing this is because it pre-dates
pathlib, but regardless, and unfortunately, their built-in path
library is a bit buggy. One type of bug is that it can't always
compare strings with it's Path objects properly, throwing errors like
those seen in:

    https://github.com/joke2k/django-environ/issues/86

So, to fix this: Simply don't use the built-in pathlib-alike, and use
the real pathlib library instead.