feat(models): Upgrade to private s3 storage
This is pretty straightforward stuff. General idea is that
we're about to move to k8s, and so we need to start storing
things in S3. The only tricky part here is that we need
to use private buckets, which is new, and we need to use
the UUID storage from before.

Fixes: #2031