fix(client): Don't repeat open cases
This bug snuck by me. The deal is that as we scrape, the data is
changing underneath us. So if we scrape a million un-terminated cases
today, kill the script and strat it on row 1,000,001 of un-terminated
cases tomorrow, that's not where we left off, because in the meantime
a bunch of cases were marked as terminated.