tests(corpus_importer): Drop assets for factory boy tests
if we want to lose testing assets I think this is the way to go
about it.  I've dropped the two new bankruptcy assets
and added a small factory file that generates the necessary
JSON for the import.  I wrote a patch to replace
the filepath generation and to substitute our data

Ive also added a new provider to generate random
citations. I think this should let us streamline tests
for importing non model based files in the future.