feat(maine): More Maine Requests
They wanted status removed on the list
And the (ME 2022) after the case name to be
removed because it didnt fit with
their conventions.  I think this makes
sense if they are going to link to it

I added a maine flag in search results
to adjust for their requests.