Pass sort order directly to docket.html
With the addition of the default sort order option for dockets we
updated the DocketEntryFilter form to clean the order_by to insert
the users default.

This resulted in the view processing the cleaned_data['order_by']
correctly.

Unfortunately, the template does not have access to the cleaned_data,
only the initial or bound data. This meant that the template would
display ascending when the default was descending when there was no
explicit order_by given on the GET.

Fixes: #2619