Fixes: #1434 add accordion to docket
When a user looks at a docket they will sometimes want a quick context.
That context might be to see the last few entries of the docket or it
could be the initial filing, which is often the complaint.

The quick look accordion presents a view into the first entry and the
latest entries on the docket.  The latest entries are the last 4 plus
any docket entries filed on the same date `date_filed` as the 4th entry.

This means that if somebody filed 3 entries yesterday and 3 entries
today they would see the first entry and the last 6 entries.

If they filed 3 entries today and 1 entry yesterday and 2 entries the
day before they would only see today's and yesterday's entries.

Further work:
* Provide filters for quick look such as this week, current week and
  last week, last 30 days and so forth
* Store the current setting of the accordion in the session so when the
  user goes to the next page the accordion is in the same state.state