Reduce memory and database load
For small dockets it is "ok" to turn a queryset into a list.  For larger
dockets this is an issue.  At the same time we want to reduce the number
of queries that are actually performed.

We set an arbitrary limit of 25 entries in the quick look.  We get all
26 possible latest docket entries with one DB call.
We make a second DB call to get the initial docket entry
We then make our third and final DB call to get the docket entries the
pagination requires