2652 Add a citation field to dockets and RECAP documents
Add includes/citation.html which takes an object and a format string and
creates a metadata field.  This includes pure javascript for copying the
citation to the clipboard and for copying the URL to the clipboard.

Added a citations template tag.  This takes in an object and a format
string.  If there is no format string then it uses a default format
string.  The code pulls information out of the object and then formats
it in reasonable ways.

If certain data is missing the tag will clean up the format before using
it.

The format string is of the form:
	{thing1} {thing2} {thing3}
This is a standard python formatting string.

There are currently a few things that the tag knows about, more can be
added.

Future work: Create a UserProfile value that allows users to select and
potentially create their own citation formatting strings.

WARNING: The code takes the formatting string as is and returns the
results as a safe string.  At this time it is not safe to hand user
created formatting strings directly to the citation tag.