Put "never" as the oldest item in a docket alert list
NULLs are sorted in magic ways by RDBMS.  Often you get the correct
answer.  When sorting by date it is more difficult, does a NULL date
mean that it is so new that it doesn't have a date or so old that it
doesn't have a date?

Django added a modifier to the `desc()` and `asc()` methods of an `F`
filter.

We updated the code to use an `F()` filter so we can explicitly say
where we want NULLS to sort in the Docket Alert List.

Fixes: #2648