When adding DocketEntries the Court Factory will sometime abort
The Court Factory randomly creates a `position`.  There is a small
chance of a collision with existing courts.  Not a serious issue when
only adding a "few" courts.

When adding DocketEntry's it is not uncommong to add thousands of
DocketEntry's.  This increases the chances of a collision.

This adds two new factories for creating DocketEntry's.  The first will
create and the new objects to existing Docket's.  The second allows for
creating DocketEntry's for a particular docket.