Fix make_dev_data --make-objects 104
The CourtFactory used Faker to create a position number.  There is a
unique constrant on the court position number.  Since this was a random
number there was a chance that there would be a collision in the
position space.

Since this is not true random numbers, this collison was happening many
times.

Object 104 creates DocketEntry's with parents, those parents include a
Docket and the Docket includes a Court.  This means that when creating
enough DocketEntry's for testing, say 5000+ it would also create 5000
Dockets and 5000 Courts

This increased the probability of a collision.

By overriding the _create() method of the CourtFactory we are able to
detect those position collisions.  When that happens we get a new random
position value and try again.  After 3 attempts we give up and raise the
exception.

In addition, there were a couple of places where we were storing a class
name rather than evaluating the class to get a string.  These are
corrected as well.