Remove copy to clipboard. Remove unused fmt
Per code review, we removed the javascript that copied the citation from
the page to the clipboard.  Now the user just needs to click on the
citation and then copy it.  Not an UX I expect but it matches the rest
of the site.

Per code review.  We are not currently using the fmt paramater on the
include of citation. Because of this it should be removed as dead code.

Issue #331 is requesting the ability for users to create or select the
format of citations, the fmt parameter was in anticipation of that use.