Add previous/next navigation buttons to the filter bar.
Currently dockets have upto 200 entries on them.  We have navigation
buttons at the very bottom that allow navigation.  Many people use
scrolling to move through a page and a 200 item docket can be time
consuming.

We add Previous and Next buttons to the filter bar to allow people to
navigate more easily when at the top of the page.

During testing it became apparent that the Bootstrap XS version was
pretty ugly, per CL design specifications.  Code was added to
override.css to make things look better.

When testing at a screen width of 360px it again looked horrible.
Having things look right at XS width and at 360 wasn't possible.

To allow for another breakpoint a series of classes were added to
override.css that provide cl-col-xxs- and cl-col-xs- as well as some
auxilary classes.  These allow for that extra breakpoint.

the cl-col-xs- are required because the col-xs- classes are always "on"
and we needed a method to override them.

We also added some specific cl-pull classes to allow different alignment
based on screen size.

Fixes: #2670