Create tight-input CSS class, remove cl- namespace
As there is no collision with xs-pull-{left,right} the cl- was removed
as a prefix.  There is a collision with form-group so that was kept.

We were modifying the padding of the inputs in order to fit on the page.
Instead of doing it based on an anchor, we are now using a class.