2694 Search pagination broken
With the new pagination code the include template commands inserted the
"only" keyword.  This strips the context of all variables that are not
explicitly passed to the included file.

One of the variables that was stripped was the get_string.  This MUST be
there in order for the parameters of the get to propigate within the
pagination.