Revamp Citation Lookup Tool
The citation lookup tool was designed to easily convert reporter,
volume, page into a citation URL which then redirects to the correct
item, if found.

We have eyecite, which can parse a text and extract a citation.  The
code was revamped to provide a single field which is capable of takeing
a citation.

EyeCite then parses that and turns it into a reporter, volume, page and
redirects to the correct item, if found.

This appears to be much cleaner, it allows for cut and paste options,
and in general looks like it is a better user experience.

All of the `/c/` lookup URLs still work the same way and the Citation
Lookup Tool home page still has a list of reporters that can be browsed.

Fixes: #2735