refactor(deploy): replace Makefile with `build-push-action`
With the Makefile radically simplified by #3260, the work that it
does can easily be done by a `docker/build-push-action` step in the
deployment workflow. Beyond than simplicity, the use of the action
makes build caching available and will trim a decent amount of time
from the build job.

In this change, the method of referring to the image is changed:
instead of a tag (e.g. `c29346b-prod`), the digest of the uploaded
image is used. The digest is an unambiguous and immutable reference
to an image, whereas a tag can change and cause surprises. For this
reason, it's recommended for deployments (see https://cloud.google.com/kubernetes-engine/docs/concepts/about-container-images).

Also in this change is the use of `docker/metadata-action`. It is
configured to produce the same tags as before (`c29346b-prod`), but
it is capable of doing much more. A future direction is to replace
the `sha` tag with either a build date or build id (`$RUN_ID-$RUN_ATTEMPT`).
For now we only introduce `labels`, which add the creation date and
full sha SHA to the built image.