feature(javascript): automate Javascript compilation
In the absense of an asset build process, compilation and
compression of Javascript must be done manually, during
development, and the output checked into the repository.
This impedes development, clutters reviews, and is an
invitation for mismatches between source and compiled
assets.

Replace the manual compiled assets with a phase in the
docker build and ad-hoc webpack server with a Docker
container, `cl-webpack`, that watches and recompiles
assets in development.