fix(security): amend CSP for media-src
CSP reports some violations where a `data:` source is
being used for media. While we have not been able to
identify the source, it is a good bet that this is
caused by a browser extension - that it is not a
manifestation of unexpected behavior in CourtListener.

Nonetheless, allow `data:`, because it is relatively
benign: it is ineffective data exfiltration and will
remove a significant cause of nuisance reports.

Fixes: #2899