chore(javascript): finish migration to Webpack 5
Webpack was upgraded to version 5 in 8256590, but has not
been used to generate Javascript. Webpack 5 includes quite
a few changes and, unfortunately, generates different
compiled output.

Notably, it strips and extracts license-like comments into
a `LICENSE.txt` file. Handle this by using a `TerserPlugin`
with a custom configuration (which is, evidently, the
recommended approach). Another difference was was source-map
generation. Seeing that existing files are essentially empty,
remove source-maps for the time being. Sadly, nothing can be
done about the different order of concatenation.

Webpack 5 Release Notes:
https://webpack.js.org/blog/2020-10-10-webpack-5-release/