fix(models): Block the J&J talcum powder case
Unfortunately, somebody is uploading every single doc in this case,
and each time they do, our poor server has to pull the massive party
information from the DB. Our server stalls while working on this,
and this backs up both the DB and Celery, leading to memory problems
in celery and a sad, slow DB.

The fix? For now (I hope), a hack to block the case from pulling party
information.