fix(citations): fix encoding of annotated plaintext opinions
Eyecite’s `annotate_citations` has an implicit requirement:
the content type of `source_text` must be compatible with
The `annotations` interpolated into it. In practice, `source_text`
should be an HTML fragment.

Here, the output of `annotate_citations` is treated as HTML,
but the input `source_text` is never encoded. After annotation,
there is no place where encoding could be done without also
encoding the annotations themselves; `opinion.source_text`
should be encoded (and `<pre>` wrapped, for good measure) as HTML
on input to Eyecite.

Fixes: #1491